fromfiled

+++ app/Http/Controllers/ListingController.php
@@ -64,6 +64,11 @@ class ListingController extends Controller
      //Update listings to database
      public function update(Request $request, Listing $listing)
      {
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
          $formField = $request->validate([
              'title' => 'required',
              'company' => ['required'],
@@ -85,14 +90,18 @@ class ListingController extends Controller
 
      //Delete listing
      public function destroy(Listing $listing) {
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
-
         return redirect('/')->with('message', 'Listing delete successfully!');
      }
 
      //manage listings
      public function manage(Listing $listing) {
-        return view('listings.manage');
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
      }
 }
